+++ common/src/main/java/software/bluelib/api/entity/variant/IVariantEntity.java
@@ -15,11 +15,13 @@ import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Entity;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import software.bluelib.api.net.NetworkRegistry;
 import software.bluelib.api.utils.logging.BaseLogLevel;
 import software.bluelib.api.utils.logging.BaseLogger;
 import software.bluelib.api.utils.variant.ParameterUtils;
 import software.bluelib.entity.variant.IVariantAccessor;
 import software.bluelib.internal.BlueTranslation;
+import software.bluelib.net.messages.client.variant.SetVariantPacket;
 
 @SuppressWarnings("unused")
 public interface IVariantEntity<T extends Entity> {
@@ -56,6 +58,9 @@ public interface IVariantEntity<T extends Entity> {
 
 	default void setVariantName(@NotNull String pVariantName) {
 		T entity = getEntity();
+		int id = entity.getId();
+		SetVariantPacket packet = new SetVariantPacket(id, pVariantName);
+		NetworkRegistry.sendToAllPlayers(entity.getServer(), packet);
 		((IVariantAccessor) entity).setEntityVariantName(pVariantName);
 	}
 }