+++ common/src/main/java/software/bluelib/internal/registry/BlueNetworkRegistry.java
@@ -15,6 +15,7 @@ import software.bluelib.api.net.PacketProvider;
 import software.bluelib.net.PacketRegisterInfo;
 import software.bluelib.net.messages.client.OpenLoggerPacket;
 import software.bluelib.net.messages.client.loader.*;
+import software.bluelib.net.messages.client.variant.SetVariantPacket;
 import software.bluelib.net.messages.server.TestPacket;
 import software.bluelib.net.serverHandling.TestPacketHandler;
 
@@ -33,6 +34,8 @@ public class BlueNetworkRegistry implements PacketProvider {
 	@Override
 	public @NotNull List<PacketRegisterInfo<?>> getS2CPackets() {
 		List<PacketRegisterInfo<?>> list = new ArrayList<>();
+
+		list.add(new PacketRegisterInfo<>(SetVariantPacket.ID, SetVariantPacket::decode));
 
 		list.add(new PacketRegisterInfo<>(OpenLoggerPacket.ID, OpenLoggerPacket::decode));