+++ common/src/main/java/software/bluelib/loader/animatable/base/AnimatableManager.java
@@ -17,7 +17,7 @@ import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import software.bluelib.loader.animation.AnimationController;
-import software.bluelib.loader.animation.bone.BoneSnapshot;
+import software.bluelib.loader.animation.keyframe.BoneFrame;
 import software.bluelib.loader.cache.controller.ControllerCache;
 import software.bluelib.loader.controller.ControllerManager;
 import software.bluelib.loader.geckolib.constant.dataticket.DataTicket;
@@ -25,7 +25,7 @@ import software.bluelib.loader.geckolib.constant.dataticket.DataTicket;
 public class AnimatableManager<T extends BlueAnimatable> {
 
 	@NotNull
-	private final Map<String, BoneSnapshot> boneSnapshotCollection = new Object2ObjectOpenHashMap<>();
+	private final Map<String, BoneFrame> boneSnapshotCollection = new Object2ObjectOpenHashMap<>();
 	@NotNull
 	private final Map<String, AnimationController<T>> animationControllers;
 	@Nullable
@@ -60,7 +60,7 @@ public class AnimatableManager<T extends BlueAnimatable> {
 	}
 
 	@NotNull
-	public Map<String, BoneSnapshot> getBoneSnapshotCollection() {
+	public Map<String, BoneFrame> getBoneSnapshotCollection() {
 		return this.boneSnapshotCollection;
 	}