+++ common/src/main/java/software/bluelib/loader/animatable/base/ContextAwareAnimatableManager.java
@@ -11,7 +11,7 @@ import java.util.Map;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import software.bluelib.loader.animation.AnimationController;
-import software.bluelib.loader.animation.bone.BoneSnapshot;
+import software.bluelib.loader.animation.keyframe.BoneFrame;
 import software.bluelib.loader.geckolib.constant.dataticket.DataTicket;
 
 public abstract class ContextAwareAnimatableManager<T extends BlueAnimatable, C> extends AnimatableManager<T> {
@@ -48,7 +48,7 @@ public abstract class ContextAwareAnimatableManager<T extends BlueAnimatable, C>
 		return getManagerForContext(getCurrentContext()).getAnimationControllers();
 	}
 
-	public @NotNull Map<String, BoneSnapshot> getBoneSnapshotCollection() {
+	public @NotNull Map<String, BoneFrame> getBoneSnapshotCollection() {
 		return getManagerForContext(getCurrentContext()).getBoneSnapshotCollection();
 	}