+++ common/src/main/java/software/bluelib/loader/cache/animations/AnimationCache.java
@@ -31,7 +31,7 @@ public record AnimationCache(
 
 	@NotNull
 	public static AnimationCache generateWaitAnimation(double pLength) {
-		return new AnimationCache(Animation.Stage.WAIT, pLength, LoopType.PLAY_ONCE, new BoneAnimationCache[0],
+		return new AnimationCache(Animation.Frame.WAIT, pLength, LoopType.PLAY_ONCE, new BoneAnimationCache[0],
 				new KeyframeLibraryCache(new SoundKeyframeData[0], new ParticleKeyframeData[0], new CustomInstructionKeyframeData[0]));
 	}